requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
coloredlogs==15.0.1
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-redis==5.2.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-tracking==1.5.0
drf-yasg==1.20.0
environs==9.5.0
factory-boy==3.2.1
Faker==13.7.0
gunicorn==20.1.0
humanfriendly==10.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
JPype1==1.4.0
lxml==4.8.0
MarkupSafe==2.1.1
marshmallow==3.16.0
numpy==1.22.4
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.4
python-stdnum==1.17
pytz==2022.1
redis==4.3.1
regex==2022.4.24
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.11.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
